Add Dockerfile and instructions for PAPYRUS-2.1.0
This is a bit more disruptive change than usual. To avoid having in the
future with Ubuntu stopping distributing the pinned version of
`openjdk-8-jdk`, instead use an official OpenJDK image as base, and pin
the SHA-256 of that image for the builder.

Additionally, modify the README.md to get all the information in a
single `docker run` and `docker cp`.