Add CD to bitcoinj-thin
Added GitHub action to build and publish the bitcoinj-thin files.
Also, the Dockerfile used was based on a deprecated version of debian,
so I updated it to use ubuntu with java 8. Please note that this is a
workaround to make things work for now, but we should be looking at
making bitcoinj-thin buildable with more recent versions of java.